Update UI to have data randomizer on navbar as well as move buttons to their own view.
